% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as.data.frame.lctmc.sim}
\alias{as.data.frame.lctmc.sim}
\title{Data tidying for the simulation output}
\usage{
\method{as.data.frame}{lctmc.sim}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A custom class list object obtained from the \code{simulate_LCTMC()} function. It should be of class 'lctmc.sim'}

\item{row.names, optional}{Ignored for objects of class 'lctmc.sim'}

\item{...}{The following argument can be specified
\describe{
\item{type}{a character scalar of either "obs", "exact", or "both". If left unspecified, then defaults to 'both'}
\item{id}{a character scalar of ID numbers. If specified, then the resulting data frame only contains the ID numbers listed in this vector. \cr
If left unspecified, then all ID numbers found in \code{x} are used}
}}
}
\value{
if \code{type = "both"} then a list object with \strong{two} elements is returned:
\enumerate{
\item \strong{obs} this data contains the disease status at some pre-defined observed times. This essentially means that all outcomes are interval-censored.
\item \strong{exact} this data contains the exact time of disease status transitioning to a different state (i.e., non-censored data)
}
If \code{type} is equal to either "obs" or "exact" then a data.frame object is returned.
The data frames corresponds to the either censored data or exactly-observed data
}
\description{
Converts the output of \code{\link[=simulate_LCTMC]{simulate_LCTMC()}} to a "data.frame" object
}
\examples{
# set seed
set.seed(456)

# simulate
d = LCTMC.simulate::simulate_LCTMC(
  N.indiv = 10,
  N.obs_times = 10,

  max.obs_times = 50,
  fix.obs_times = FALSE,

  true_param = LCTMC.simulate::gen_true_param(K_class = 3, M_state = 2),
  alpha.include = TRUE,
  beta.include = TRUE,

  K = 3,
  M = 2,
  p1 = 2,
  p2 = 2,

  initS_p = c(0.5, 0.5),
  death = NULL
)

# coerce to data.frame ~ S3 method for 'lctmc.sim' objects
head(as.data.frame(x = d, type = "obs"))
head(as.data.frame(x = d, type = "exact"))

# leaving `type` unspecified
sim_df_list = as.data.frame(x = d)
class(sim_df_list)
names(sim_df_list)

# specify an ID number
as.data.frame(x = d, id = c("EA000"))
}
\seealso{
\code{\link[=simulate_LCTMC]{simulate_LCTMC()}}
}
