% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_simulate_LCTMC.R
\name{check_simulate_LCTMC}
\alias{check_simulate_LCTMC}
\title{Checks \code{simulate_LCTMC()}}
\usage{
check_simulate_LCTMC(
  N.indiv = integer(),
  N.obs_times = integer(),
  max.obs_times = numeric(),
  fix.obs_times = logical(),
  true_param = list(),
  alpha.include = logical(),
  beta.include = logical(),
  K = integer(),
  M = integer(),
  p1 = integer(),
  p2 = integer(),
  initS_p = c(),
  death = integer()
)
}
\arguments{
\item{N.indiv}{See documentation for \code{simulate_LCTMC()}}

\item{N.obs_times}{See documentation for \code{simulate_LCTMC()}}

\item{max.obs_times}{See documentation for \code{simulate_LCTMC()}}

\item{fix.obs_times}{See documentation for \code{simulate_LCTMC()}}

\item{true_param}{See documentation for \code{simulate_LCTMC()}}

\item{alpha.include}{See documentation for \code{simulate_LCTMC()}}

\item{beta.include}{See documentation for \code{simulate_LCTMC()}}

\item{K}{See documentation for \code{simulate_LCTMC()}}

\item{M}{See documentation for \code{simulate_LCTMC()}}

\item{p1}{See documentation for \code{simulate_LCTMC()}}

\item{p2}{See documentation for \code{simulate_LCTMC()}}

\item{initS_p}{See documentation for \code{simulate_LCTMC()}}

\item{death}{See documentation for \code{simulate_LCTMC()}}
}
\value{
if checks passes without issue then function returns NULL
}
\description{
Takes the exact same arguments as \code{simulate_LCTMC(...)} and checks if there any issues with specifications.
}
\seealso{
\code{\link[=simulate_LCTMC]{simulate_LCTMC()}}
}
