% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_IDlist.R
\name{gen_IDlist}
\alias{gen_IDlist}
\title{Generates a list of ID numbers}
\usage{
gen_IDlist()
}
\value{
a character vector containing unique IDs
}
\description{
Generate a list of character class IDs to be used for the CTMC simulation.
The format is "AAxxx" where "AA" are two capital letters and "xxx" are three digit numbers (leading zeroes are allowed, e.g., FX001)
}
\examples{
\dontrun{
  IDlist = gen_IDlist()
  IDlist
}
}
