% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_obsTime.R
\name{gen_obsTime}
\alias{gen_obsTime}
\title{Generate known observation times}
\usage{
gen_obsTime(N.obs_times, min_t, max_t)
}
\arguments{
\item{N.obs_times}{number of observation to be made}

\item{min_t}{the smallest possible time value an observation can be made}

\item{max_t}{the largest possible time value an observation can be made}
}
\value{
a numeric vector in ascending order which are the observation times for the CTMC process, with zero appended as the first element to indicate time of origin.
}
\description{
One key aspect of a CTMC process is that observations are made only between the exact transition times which are unknown. \cr
This function's purpose is to generate those observation times which are independent of the times when transition occurs.
}
\note{
this function essentially simulates uniform random variables and sorting them in ascending order.
}
\examples{
\dontrun{
  # min time = 0, max time = 10. Make 10 observations
  set1 = gen_obsTime(N.obs_times = 10, min_t = 0, max_t = 10)
  set1

  # min time = 0, max time = 1000. Make 10 observations
  # notice the INCREASE in change in time, compare to set1
  gen_obsTime(N.obs_times = 10, min_t = 0, max_t = 1000)

  # min time = 0, max time = 1. Make 100 observations
  # notice the DECREASE in change in time, compare to set1
  gen_obsTime(N.obs_times = 100, min_t = 0, max_t = 1)
}
}
