% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_true_param.R
\name{gen_true_param}
\alias{gen_true_param}
\title{Generates true model parameters}
\usage{
gen_true_param(
  K_class = integer(),
  M_state = integer(),
  pi.Z1 = NULL,
  pi.Z2 = NULL,
  r0.Z1 = NULL,
  r0.Z2 = NULL,
  r0.Z3 = NULL,
  beta.Z1 = NULL,
  beta.Z2 = NULL,
  beta.Z3 = NULL
)
}
\arguments{
\item{K_class}{an integer scalar. The number of latent classes (currently only supports 3 class). \cr Default is 3}

\item{M_state}{an integer scalar. The number CTMC states (currently only supports 2 or 3 states). \cr Default is 2}

\item{pi.Z1, pi.Z2}{a list object containing 3 elements for the multinomial logistic model: \code{alpha0}, \code{alpha1}, \code{alpha2}. \cr
\code{pi.Z1} will be used for class1's coefficient. \code{pi.Z2} for class2. And class3 is the referent. \cr
For example: \code{pi.Z1 = list(alpha0 = 0.7123, alpha1 = 0.8678, alpha2 = 1.1234)}}

\item{r0.Z1, r0.Z2, r0.Z3}{a list object containing 2 elements for a two-state model, and 6 elements for a 3-state model. \cr
These parameters are used to specified the intercept terms of the transition rates (see sections below for the model formulation).
Each list object is assigned to the respective latent class.
For example: \code{r0.Z3 = list(q12 = 1.05, q21 = 1.10)}}

\item{beta.Z1, beta.Z2, beta.Z3}{a list object containing 2 elements for a two-state model, and 6 elements for a 3-state model. \cr
These parameters are used to specified the coefficient terms of the transition rates (see sections below for the model formulation).
Each list object is assigned to the respective latent class. \cr
For example: \code{beta.Z1 = list(q12 = c(1.32, -0.29), q21 = c(-0.45, 0.65))}}
}
\value{
A nested list object containing three elements: \code{r0}, \code{beta}, and \code{pi}
\describe{
\item{r0}{a nested list, one element for each \eqn{k} class, containing the \code{r0} component of the formula: \cr
\deqn{
          q_{rs(k)} = r0_{rs(k)} \cdot exp(\beta_{rs(k)}X)
        }}
\item{beta}{a nested list, one element for each \eqn{k} class, containing the \code{beta} component of the formula: \cr
\deqn{
          q_{rs(k)} = r0_{rs(k)} \cdot exp(\beta_{rs(k)}X)
        }}
\item{pi}{a nested list of \eqn{K-1} elements because the last class is the referent group. These parameters are the coefficients of the multinomial logistic model:
\deqn{
          log(\frac{\pi_{(k)}}{\pi_{K}}) = \alpha_{(k)}W
        }}
}
}
\description{
Generates the true model parameters values for simulating synthetic data.
}
\note{
For the beta arguments, each element within the list is a numeric vector of length 2. This is because we allow the simulation to
take up to two covariates for the CTMC model. If these model parameter arguments are set to \code{NULL} then the function will internally supply some default values.
}
\examples{
# 3 latent class , general 2x2 case
gen_true_param(K_class = 3, M_state = 2)

# 3 latent class , special 3x3 case , where q13, q31, q32 are equal to zero
gen_true_param(K_class = 3, M_state = 3)

# an example where user specifies their own parameter values
# note: the ones not specified take default values
gen_true_param(
  K_class = 3, M_state = 3,
  pi.Z1 = list(alph0 = 1, alpha1 = 1, alpha2 = 1),
  r0.Z2 = list(q12 = 9, q21 = 9, q21 = 9, q23 = 9, q31 = 9, q32 = 9),
  beta.Z3 = list(q12 = c(0,1), q21 = c(0,1), q21 = c(0,1), q23 = c(0,1), q31 = c(0,1), q32 = c(0,1))
)
}
