% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.lctmc.sim}
\alias{plot.lctmc.sim}
\title{Visualizes simulated data}
\usage{
\method{plot}{lctmc.sim}(x, ...)
}
\arguments{
\item{x}{A custom class list object obtained from \code{simulate_LCTMC()}. The custom class is called 'lctmc.sim'}

\item{...}{The following variable should be specified within \code{...}
\describe{
\item{id}{A character scalar to indicate which person (ID number) should be plotted}
}}
}
\value{
This function generates a plot and returns \code{NULL}
}
\description{
This function visualizes the simulated longitudinal data for the given ID number
}
\examples{
# set seed
set.seed(456)

# simulate
d = LCTMC.simulate::simulate_LCTMC(
  N.indiv = 5,
  N.obs_times = 5,

  max.obs_times = 9,
  fix.obs_times = FALSE,

  true_param = LCTMC.simulate::gen_true_param(K_class = 3, M_state = 2),
  alpha.include = TRUE,
  beta.include = TRUE,

  K = 3,
  M = 2,
  p1 = 2,
  p2 = 2,

  initS_p = c(0.5, 0.5),
  death = NULL,
  sojourn = 'gamma'
)

# plot ~ S3 method for 'lctmc.sim' objects
plot(x = d, id = "BA000")
}
\seealso{
\code{\link[=simulate_LCTMC]{simulate_LCTMC()}}, \code{\link[=as.data.frame.lctmc.sim]{as.data.frame.lctmc.sim()}}
}
