% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transitions.R
\name{plot_transitions}
\alias{plot_transitions}
\title{Simulate a data from a latent CTMC model}
\usage{
plot_transitions(df, id)
}
\arguments{
\item{df}{a list object obtained from \code{convert_sim_data_2df(...)}. This list object should hold two data frame objects \code{obs} and \code{exact}. \cr
See \code{?convert_sim_data_2df} for more details}

\item{id}{a character scalar to indicate which person (ID number) should be plotted}
}
\value{
This function generates a plot and returns \code{NULL}
}
\description{
Given some underlying parameter values that dictate the latent class assignment process, the CTMC process, and observation times.
simulates a latent CTMC process and outputs the result to a list object.
}
\examples{
# set seed
set.seed(456)

# simulate
d = LCTMC.simulate::simulate_LCTMC(
  N.indiv = 5,
  N.obs_times = 5,

  max.obs_times = 9,
  fix.obs_times = FALSE,

  true_param = LCTMC.simulate::gen_true_param(K_class = 3, M_state = 2),
  alpha.include = TRUE,
  beta.include = TRUE,

  K = 3,
  M = 2,
  p1 = 2,
  p2 = 2,

  initS_p = c(0.5, 0.5),
  death = NULL
)

# convert to data.frames
my_df = LCTMC.simulate::convert_sim_data_2df(my_list = d$sim_data, type = "both")

# plot
LCTMC.simulate::plot_transitions(df = my_df, id = "EA000")
}
\seealso{
\code{\link[=simulate_LCTMC]{simulate_LCTMC()}}, \code{\link[=convert_sim_data_2df]{convert_sim_data_2df()}}
}
