% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sim_data_2df.R
\name{convert_sim_data_2df}
\alias{convert_sim_data_2df}
\title{Data tidying for the simulation output}
\usage{
convert_sim_data_2df(my_list, type)
}
\arguments{
\item{my_list}{a list object obtained from the \code{simulate_LCTMC()} function}

\item{type}{a character scalar of either "obs", "exact", or "both"}
}
\value{
if \code{type = "both"} then a a list object with \strong{2} elements is returned:
\enumerate{
\item \strong{obs} this data contains the disease status at some pre-defined observed times. This essentially means that all outcomes are interval-censored.
\item \strong{exact} this data contains the exact time of disease status transitioning to a different state (i.e., non-censored data)
}
If \code{type} is equal to either "obs" or "exact" then a data.frame object is returned. The data frames corresponds to the either censored data or exactly-observed data
}
\description{
Converts the output of \code{\link[=simulate_LCTMC]{simulate_LCTMC()}} to a "data.frame" object
}
\examples{
# set seed
set.seed(123)

# simulate
d = LCTMC.simulate::simulate_LCTMC(
  N.indiv = 10,
  N.obs_times = 10,

  max.obs_times = 50,
  fix.obs_times = FALSE,

  true_param = LCTMC.simulate::gen_true_param(K_class = 3, M_state = 2),
  alpha.include = TRUE,
  beta.include = TRUE,

  K = 3,
  M = 2,
  p1 = 2,
  p2 = 2,

  initS_p = c(0.5, 0.5),
  death = NULL
)

# convert to data.frame
head(LCTMC.simulate::convert_sim_data_2df(my_list = d$sim_data, type = "obs"))
head(LCTMC.simulate::convert_sim_data_2df(my_list = d$sim_data, type = "exact"))
}
\seealso{
\code{\link[=simulate_LCTMC]{simulate_LCTMC()}}
}
