% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_transition.R
\name{gen_transition}
\alias{gen_transition}
\title{Simulates a transition with exponentially distributed sojourn time}
\usage{
gen_transition(from_state, Q, M_state)
}
\arguments{
\item{from_state}{the current state to transition out of}

\item{Q}{transition rate matrix obtained from \code{gen_Qmat()}}

\item{M_state}{number of CTMC states (currently only support 2 or 3 states)}
}
\value{
a list containing two elements
\enumerate{
\item \strong{t} a numeric scalar. The exponentially distributed sojourn time it took to transition out of \code{from_state}
\item \strong{to_state} a numeric scalar. This is the state which the random process transitioned to
\item if something went wrong then NA values are returned
}
}
\description{
Generates the process of state transitioning. This is done by following 2 steps: \cr
(1) generate the time at which transition actually occurs \cr
(2) generate which what state is transitioned to
}
\examples{
\dontrun{
  # generate some CTMC parameters
  tp = gen_true_param(K_class = 3, M_state = 3)

  # use the given CTMC parameters to generate Q matrix
  x1 = 1
  x2 = 2
  x_covariate = c(1,2)
  Q = LCTMC.simulate:::gen_Qmat(r0 = tp$r0,
                                beta = tp$beta,
                                x = x_covariate,
                                z = 1,
                                M_state = 3,
                                K_class = 3)

  # generate a transition
  gen_transition(from_state = 2, Q = Q, M_state = 3)
}
}
\seealso{
\code{\link[=gen_Qmat]{gen_Qmat()}}
}
