% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transitions.R
\name{plot_transitions}
\alias{plot_transitions}
\title{Simulate a data from a latent CTMC model}
\usage{
plot_transitions(df, id)
}
\arguments{
\item{df}{a list object obtained from \code{convert_sim_data_2df(...)}. This list object should hold two data frame objects \code{obs} and \code{exact}. \cr
See \code{?convert_sim_data_2df} for more details}

\item{id}{a character scalar to indicate which person (ID number) should be plotted}
}
\value{
This function generates a plot and returns \code{NULL}
}
\description{
Given some underlying parameter values that dictate the latent class assignment process, the CTMC process, and observation times.
simulates a latent CTMC process and outputs the result to a list object.
}
\examples{
# simulation parameters
my_N = 5
my_N.obs_times = 5
my_max.obs_times = 9
my_fix.obs_times = FALSE

my_beta.include = TRUE
my_M = 2
my_p1 = 2

my_alpha.include = TRUE
my_K = 3
my_p2 = 2

my_initS_p = c(1/2, 1/2)
my_death = NULL # for a two-stage process we do not have death state, set to NULL

set.seed(456)

# first generate true parameters
my_true_param = gen_true_param(K_class = 3, M_state = 2)

# simulate
d = simulate_LCTMC(
  N.indiv = my_N,
  N.obs_times = my_N.obs_times,

  max.obs_times = my_max.obs_times,
  fix.obs_times = my_fix.obs_times,

  true_param = my_true_param,
  alpha.include = my_alpha.include,
  beta.incldue = my_beta.include,

  K = my_K,
  M = my_M,
  p1 = my_p1,
  p2 = my_p2,

  initS_p = my_initS_p,
  death = my_death
)

# convert to data.frames
my_df = convert_sim_data_2df(my_list = d$sim_data, type = "both")

# plot
plot_transitions(df = my_df, id = "EA000")
}
